Fix: user 테스트 오류 수정 , 지역 테이블 데이터 저장

+++ Back/petPlace/src/main/java/com/minjeok4go/petplace/domain/feed/model/FeedDetailDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class FeedDetailDto {
     private Long id;
     private String content;
-    private Long userId;            // ✅ uid → userId
+    private Integer userId;            // ✅ uid → userId (타입 수정: Long → Integer)
     private String userNick;
     private String userImg;
     private Long regionId;          // ✅ rid → regionId
